requirements.txt
python-dotenv
discord
pydantic
pydantic-settings
dateparser
innertube
tinydb
